Move `pkg/beacon/group` out of `pkg/beacon`
This change will allow reusing the group
by other applications.
