+++ pkg/beacon/dkg/marshalling.go
@@ -5,8 +5,8 @@ import (
 	"math/big"
 
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
-	"github.com/keep-network/keep-core/pkg/beacon/group"
 	"github.com/keep-network/keep-core/pkg/beacon/registry/gen/pb"
+	"github.com/keep-network/keep-core/pkg/group"
 )
 
 // Marshal converts ThresholdSigner to byte array.
