+++ pkg/beacon/dkg/result/submission_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	beaconchain "github.com/keep-network/keep-core/pkg/beacon/chain"
-	"github.com/keep-network/keep-core/pkg/beacon/group"
 	"github.com/keep-network/keep-core/pkg/chain"
+	"github.com/keep-network/keep-core/pkg/group"
 )
 
 func TestSubmitDKGResult(t *testing.T) {
