+++ pkg/beacon/dkg/signer.go
@@ -5,8 +5,8 @@ import (
 
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
 	"github.com/keep-network/keep-core/pkg/altbn128"
-	"github.com/keep-network/keep-core/pkg/beacon/group"
 	"github.com/keep-network/keep-core/pkg/bls"
+	"github.com/keep-network/keep-core/pkg/group"
 )
 
 // ThresholdSigner is created from GJKR group Member when DKG protocol completed
