+++ pkg/beacon/dkg/signer_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
-	"github.com/keep-network/keep-core/pkg/beacon/group"
 	"github.com/keep-network/keep-core/pkg/bls"
+	"github.com/keep-network/keep-core/pkg/group"
 )
 
 func TestSignAndComplete(t *testing.T) {
