+++ pkg/beacon/entry/message.go
@@ -1,7 +1,7 @@
 package entry
 
 import (
-	"github.com/keep-network/keep-core/pkg/beacon/group"
+	"github.com/keep-network/keep-core/pkg/group"
 )
 
 // SignatureShareMessage is a message payload that carries the sender's
