+++ pkg/beacon/gjkr/export_test.go
@@ -4,8 +4,8 @@ import (
 	"math/big"
 
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
-	"github.com/keep-network/keep-core/pkg/beacon/group"
 	"github.com/keep-network/keep-core/pkg/crypto/ephemeral"
+	"github.com/keep-network/keep-core/pkg/group"
 )
 
 func (epkm *EphemeralPublicKeyMessage) SetSenderID(
