+++ pkg/beacon/gjkr/marshaling.go
@@ -5,8 +5,8 @@ import (
 
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
 	"github.com/keep-network/keep-core/pkg/beacon/gjkr/gen/pb"
-	"github.com/keep-network/keep-core/pkg/beacon/group"
 	"github.com/keep-network/keep-core/pkg/crypto/ephemeral"
+	"github.com/keep-network/keep-core/pkg/group"
 )
 
 // MemberIndex is represented as uint8 in gjkr. Protobuf does not have uint8
