+++ pkg/beacon/gjkr/marshaling_test.go
@@ -8,8 +8,8 @@ import (
 	fuzz "github.com/google/gofuzz"
 
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
-	"github.com/keep-network/keep-core/pkg/beacon/group"
 	"github.com/keep-network/keep-core/pkg/crypto/ephemeral"
+	"github.com/keep-network/keep-core/pkg/group"
 	"github.com/keep-network/keep-core/pkg/internal/pbutils"
 )
