+++ pkg/beacon/gjkr/protocol_accusations_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
-	"github.com/keep-network/keep-core/pkg/beacon/group"
 	"github.com/keep-network/keep-core/pkg/crypto/ephemeral"
+	"github.com/keep-network/keep-core/pkg/group"
 )
 
 // TODO Add test with many messages from accusers and many accused in the message.
