+++ pkg/beacon/gjkr/protocol_combinations_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
-	"github.com/keep-network/keep-core/pkg/beacon/group"
+	"github.com/keep-network/keep-core/pkg/group"
 )
 
 func TestCombineGroupPublicKey(t *testing.T) {
