+++ pkg/beacon/gjkr/protocol_reconstructions_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
-	"github.com/keep-network/keep-core/pkg/beacon/group"
 	"github.com/keep-network/keep-core/pkg/crypto/ephemeral"
+	"github.com/keep-network/keep-core/pkg/group"
 )
 
 func TestRevealMisbehavedMembersKeys(t *testing.T) {
