+++ pkg/beacon/gjkr/result.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
-	"github.com/keep-network/keep-core/pkg/beacon/group"
+	"github.com/keep-network/keep-core/pkg/group"
 )
 
 // Result of distributed key generation protocol.
