+++ pkg/beacon/gjkr/states.go
@@ -3,7 +3,7 @@ package gjkr
 import (
 	"context"
 
-	"github.com/keep-network/keep-core/pkg/beacon/group"
+	"github.com/keep-network/keep-core/pkg/group"
 	"github.com/keep-network/keep-core/pkg/net"
 	"github.com/keep-network/keep-core/pkg/state"
 )
