+++ pkg/internal/dkgtest/assertions.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/keep-network/keep-core/pkg/altbn128"
-	"github.com/keep-network/keep-core/pkg/beacon/group"
+	"github.com/keep-network/keep-core/pkg/group"
 	"github.com/keep-network/keep-core/pkg/internal/testutils"
 )
